+++ ChangeLog.md
@@ -3,6 +3,7 @@ ChangeLog
 
 ## unreleased:
 * Add `{{symbol}}` to get a new replacement for a symbol.
+* Add new escapement '~ ' to escape spaces.
 
 ## 0.1.0 (2019-04-13):
 * Add support for JSON.